refactor: rename hostid6/mac.go to parse.go
The hostid6 package organizes files by concern (model, application,
parsing), not by derivation kind. The mac.go name implied it owned all
MAC logic, but it only owns the strict 48-bit MAC text parser; the rest
of the MAC handling lives in hostid6.go and derive.go alongside the
other kinds. Rename to parse.go to match the package's concern-based
layout, and move the ParseMAC tests to parse_test.go to match.
