+++ internal/hostid6/parse.go
